refactor: Participant 추상화로 Player, Dealer 중복코드 제거
